Move type to global internal scope
